+++ nextstrain/cli/runner/ambient.py
@@ -35,10 +35,12 @@ Once you've installed dependencies, proceed with ``nextstrain setup ambient``.
 
 import os
 import shutil
+import sys
 from subprocess import CalledProcessError
-from typing import Iterable
-from ..types import Env, RunnerSetupStatus, RunnerTestResults, RunnerUpdateStatus
-from ..util import capture_output, exec_or_return
+from typing import Iterable, cast
+from .. import config
+from ..types import Env, RunnerModule, SetupStatus, SetupTestResults, UpdateStatus
+from ..util import capture_output, exec_or_return, runner_name
 
 
 def register_arguments(parser) -> None:
@@ -59,14 +61,14 @@ def run(opts, argv, working_volume = None, extra_env: Env = {}, cpus: int = None
     return exec_or_return(argv, extra_env)
 
 
-def setup(dry_run: bool = False, force: bool = False) -> RunnerSetupStatus:
+def setup(dry_run: bool = False, force: bool = False) -> SetupStatus:
     """
     Not supported.
     """
     return None
 
 
-def test_setup() -> RunnerTestResults:
+def test_setup() -> SetupTestResults:
     def runnable(*argv) -> bool:
         try:
             capture_output(argv)
@@ -87,12 +89,12 @@ def test_setup() -> RunnerTestResults:
 
 def set_default_config() -> None:
     """
-    No-op.
+    Sets ``core.runner`` to this runner's name (``ambient``).
     """
-    pass
+    config.set("core", "runner", runner_name(cast(RunnerModule, sys.modules[__name__])))
 
 
-def update() -> RunnerUpdateStatus:
+def update() -> UpdateStatus:
     """
     Not supported.  Updating the ambient environment isn't reasonably possible.
     """