+++ nextstrain/cli/runner/aws_batch/__init__.py
@@ -80,16 +80,17 @@ defaults set by `config file variables`_.
 import botocore.exceptions
 import os
 import shlex
+import sys
 from datetime import datetime
 from pathlib import Path
 from signal import signal, Signals, SIGINT
 from sys import exit, stdin
 from textwrap import dedent
 from time import sleep, time
-from typing import Iterable, Optional
+from typing import Iterable, Optional, cast
 from uuid import uuid4
-from ...types import Env, RunnerSetupStatus, RunnerTestResults, RunnerUpdateStatus
-from ...util import colored, prose_list, warn
+from ...types import Env, RunnerModule, SetupStatus, SetupTestResults, UpdateStatus
+from ...util import colored, prose_list, runner_name, warn
 from ... import config
 from .. import docker
 from . import jobs, s3
@@ -559,14 +560,14 @@ def generate_run_id() -> str:
     return str(uuid4())
 
 
-def setup(dry_run: bool = False, force: bool = False) -> RunnerSetupStatus:
+def setup(dry_run: bool = False, force: bool = False) -> SetupStatus:
     """
     Not supported.
     """
     return None
 
 
-def test_setup() -> RunnerTestResults:
+def test_setup() -> SetupTestResults:
     """
     Check that necessary AWS resources exist.
     """
@@ -582,12 +583,12 @@ def test_setup() -> RunnerTestResults:
 
 def set_default_config() -> None:
     """
-    No-op.
+    Sets ``core.runner`` to this runner's name (``aws-batch``).
     """
-    pass
+    config.set("core", "runner", runner_name(cast(RunnerModule, sys.modules[__name__])))
 
 
-def update() -> RunnerUpdateStatus:
+def update() -> UpdateStatus:
     """
     Not supported.  Updating the AWS Batch runtime isn't meaningful.
     """