+++ nextstrain/cli/runner/singularity.py
@@ -85,16 +85,17 @@ import os
 import re
 import shutil
 import subprocess
+import sys
 from functools import lru_cache
 from packaging.version import Version, InvalidVersion
 from pathlib import Path
-from typing import Iterable, List, Optional
+from typing import Iterable, List, Optional, cast
 from urllib.parse import urlsplit
 from .. import config
 from ..errors import UserError
 from ..paths import RUNTIMES
-from ..types import Env, RunnerSetupStatus, RunnerTestResults, RunnerUpdateStatus
-from ..util import capture_output, colored, exec_or_return, split_image_name, warn
+from ..types import Env, RunnerModule, SetupStatus, SetupTestResults, UpdateStatus
+from ..util import capture_output, colored, exec_or_return, runner_name, split_image_name, warn
 from . import docker
 
 flatten = itertools.chain.from_iterable
@@ -273,7 +274,7 @@ def run(opts, argv, working_volume = None, extra_env: Env = {}, cpus: int = None
     ], extra_env)
 
 
-def setup(dry_run: bool = False, force: bool = False) -> RunnerSetupStatus:
+def setup(dry_run: bool = False, force: bool = False) -> SetupStatus:
     if not setup_image(dry_run, force):
         return False
 
@@ -309,7 +310,7 @@ def setup_image(dry_run: bool = False, force: bool = False) -> bool:
     return True
 
 
-def test_setup() -> RunnerTestResults:
+def test_setup() -> SetupTestResults:
     def test_run():
         try:
             capture_output([
@@ -345,13 +346,16 @@ def test_setup() -> RunnerTestResults:
 
 def set_default_config() -> None:
     """
+    Sets ``core.runner`` to this runner's name (``singularity``).
+
     Sets ``singularity.image``, if it isn't already set, to the latest
     ``build-*`` image.
     """
+    config.set("core", "runner", runner_name(cast(RunnerModule, sys.modules[__name__])))
     config.setdefault("singularity", "image", latest_build_image(DEFAULT_IMAGE))
 
 
-def update() -> RunnerUpdateStatus:
+def update() -> UpdateStatus:
     """
     Download and convert the latest Docker runtime image into a local
     Singularity image.
@@ -361,7 +365,7 @@ def update() -> RunnerUpdateStatus:
     return _update()
 
 
-def _update(dry_run: bool = False) -> RunnerUpdateStatus:
+def _update(dry_run: bool = False) -> UpdateStatus:
     current_image = DEFAULT_IMAGE
     latest_image  = latest_build_image(current_image)